--- a/README.md
+++ b/README.md
@@ -49,3 +49,4 @@ Give a ⭐️ if you like this project!
 
 ## 📝 License
 
+This project is [MIT](./MIT.md) licensed.