--- a/data.sql
+++ b/data.sql
@@ -82,3 +82,10 @@ insert into  visits(vet_id, animal_id, date_of_visit) values(2, 9, '2020-02-27')
 insert into  visits(vet_id, animal_id, date_of_visit) values(2, 9, '2020-08-03');
 insert into  visits(vet_id, animal_id, date_of_visit) values(3, 10, '2020-05-24');
 insert into visits(vet_id, animal_id, date_of_visit) values(1, 10, '2021-01-11');
+
+
+-- Insert visitors data 
+
+INSERT INTO visits (animal_id, vet_id, date_of_visit) SELECT * FROM (SELECT id FROM animals) animal_ids, (SELECT id FROM vets) vets_ids, generate_series('1980-01-01'::timestamp, '2021-01-01', '4 hours') visit_timestamp;
+
+INSERT INTO "Owners" (full_name, email) SELECT 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';