--- a/schema.sql
+++ b/schema.sql
@@ -29,3 +29,27 @@ create table animals(
 create table visits (vet_id int references vets(id),animal_id int references animals(id),date_of_visit date;
 );
 create table specializations (vet_id int references vets(id),species_id int references species(id));
+
+
+
+alter table animals ADD owner_id INTEGER REFERENCES owners(id)
+
+--  create owners table
+create table owners (
+	id SERIAL PRIMARY KEY,
+@@ -61,4 +58,8 @@ CREATE TABLE visits (
+	vet_id INTEGER REFERENCES vets(id),
+	animal_id INTEGER REFERENCES animals(id),
+	date_of_visit DATE
+);
+-- add email field on owners table
+alter table owners add COLUMN email varchar(120);
+
+  -- performance audit
+
+EXPLAIN ANALYZE SELECT COUNT(*) FROM visits WHERE animal_id = 4
+-- the above query returns an execution time of 4380 ms
+CREATE INDEX visits_animal_id_idx ON visits (animal_id);
+-- after creating the index, the query execution time is Execution Time: 770.499 ms
+
+