--- a/schema.sql
+++ b/schema.sql
@@ -56,9 +56,17 @@ CREATE INDEX visits_animal_id_idx ON visits (animal_id);
 
 
 EXPLAIN ANALYZE SELECT * FROM visits WHERE vet_id = 2;
--- the above query returns an execution time of 1060.80 ms
+  -- the above query returns an execution time of 1060.80 ms
 
 CREATE INDEX visits_vet_id_idx ON visits (vet_id);
--- after creating the index, the query  Execution Time: 799.856 ms
- ms
+  -- after creating the index, the query  Execution Time: 799.856 ms
+
+
+EXPLAIN ANALYZE SELECT * FROM owners WHERE email = 'owner_18327@mail.com';
+-- the above query returns an execution time of 4886.621 ms
+
+CREATE INDEX owners_email_idx ON owners (email);
+-- after creating the index, the query eecution Time: 0.112 ms
+
+