### Projects achievements 

We were able to optimize the queries below for better performance and scalability 

**Task one** 
- SELECT COUNT(*) FROM visits where animal_id = 4;
          -- the above query returns an execution time of **4380** **ms**
          after adding the following  index query =>    _CREATE INDEX visits_animal_id_idx ON visits (animal_id);_
           the query execution time is Execution Time: **770.499 ms**
_Before adding the index_ 

![1analyse](https://user-images.githubusercontent.com/100305514/191068570-fd8d13f9-776b-453b-9e35-7d3a8a85ad1a.png)

_After adding the index_ 

![11Ana](https://user-images.githubusercontent.com/100305514/191069022-f2a0bae7-aa83-4b01-a4cf-335d6d67784a.png)

**Task Two** 
- SELECT * FROM visits where vet_id = 2;
  -- the above query returns an execution time of **1060.80 ms**
  -- after creating the index, the query  Execution Time of **: 799.856 ms**
 
_After adding the index_ 
![2analyse](https://user-images.githubusercontent.com/100305514/191070483-7b4675a4-bf88-4a80-801f-2496e57b7e7f.png)

_Before adding the index_ 
![22aNA](https://user-images.githubusercontent.com/100305514/191071196-ae4e2512-71e3-4c38-8ecf-33cb08e3c8e7.png)

**Task Three** 
- SELECT * FROM owners where email = 'owner_18327@mail.com'; 
-   -- the above query returns an execution time of **1060.80 ms**
  -- after creating the index, the query  Execution Time: **0.112 ms**


_Before adding the index_ 
![A4](https://user-images.githubusercontent.com/100305514/191072201-21ca77e0-2e1e-418c-b2ba-2c9dfb876abc.png)

 _After adding the index_ 
![AA4](https://user-images.githubusercontent.com/100305514/191072142-927a15e4-e736-4086-869c-7fdaa9ea1235.png)